+++ docs/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - #92 List result variables on final report as order in sample view
+- #89 Show patient name in analysis reports dashboard
+- #88 Result interpretation comment showing twice
 - #87 Remove preliminary comment
 - #85 Add Result Variable's unit in final report
 - #75 Make Department field required based on Client in Sample View